celery tasks refactor to have celery related code and config in /backend/tasks folder
