+++ backend/tasks/celery_app.py
@@ -44,8 +44,8 @@ celery_app = Celery(
     broker=broker_url,
     backend=result_backend,
     include=[
-        "backend.routes.user.code_validation",
-        "backend.games.simulation_task",
+        "backend.tasks.validation_task",
+        "backend.tasks.simulation_task",
     ],
 )
 
@@ -60,7 +60,7 @@ celery_app.conf.update(
     # Fresh process per task: untrusted agent code can monkeypatch games.* or
     # leak module state — the process boundary is the isolation guarantee.
     # If this is ever relaxed, the simulation task's per-child DB engine
-    # assumption breaks too (see backend/games/simulation_task.py).
+    # assumption breaks too (see backend/tasks/simulation_task.py).
     worker_max_tasks_per_child=1,
     worker_prefetch_multiplier=1,
     # No task uses rate limits; skipping the per-task token-bucket bookkeeping
