+++ docker-compose.yml
@@ -69,7 +69,7 @@ services:
     build:
       context: .
       dockerfile: ./backend/Dockerfile
-    command: celery -A backend.celery_app worker -Q validation -n validation@%h --loglevel=info --concurrency=4
+    command: celery -A backend.tasks.celery_app worker -Q validation -n validation@%h --loglevel=info --concurrency=4
     environment:
       - PYTHONDONTWRITEBYTECODE=1
       - PYTHONUNBUFFERED=1
@@ -86,7 +86,7 @@ services:
         condition: service_healthy
     restart: unless-stopped
     healthcheck:
-      test: ["CMD-SHELL", "celery -A backend.celery_app inspect ping -d validation@$$HOSTNAME -t 10 | grep -q pong"]
+      test: ["CMD-SHELL", "celery -A backend.tasks.celery_app inspect ping -d validation@$$HOSTNAME -t 10 | grep -q pong"]
       interval: 30s
       timeout: 15s
       retries: 3
@@ -101,7 +101,7 @@ services:
     build:
       context: .
       dockerfile: ./backend/Dockerfile
-    command: celery -A backend.celery_app worker -Q simulation -n simulation@%h --loglevel=info --concurrency=1
+    command: celery -A backend.tasks.celery_app worker -Q simulation -n simulation@%h --loglevel=info --concurrency=1
     environment:
       - PYTHONDONTWRITEBYTECODE=1
       - PYTHONUNBUFFERED=1
@@ -118,7 +118,7 @@ services:
         condition: service_healthy
     restart: unless-stopped
     healthcheck:
-      test: ["CMD-SHELL", "celery -A backend.celery_app inspect ping -d simulation@$$HOSTNAME -t 10 | grep -q pong"]
+      test: ["CMD-SHELL", "celery -A backend.tasks.celery_app inspect ping -d simulation@$$HOSTNAME -t 10 | grep -q pong"]
       interval: 30s
       timeout: 15s
       retries: 3
